requirements.txt
fastapi==0.115.0
uvicorn==0.31.1
sqlalchemy==2.0.35
pydantic==2.9.2
pytest==8.3.3
python-dotenv==1.0.1
starlette==0.37.2
pyodbc==5.1.0
passlib==1.7.4
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
requests==2.32.3
pydantic-settings==2.5.2
python-multipart==0.0.12
psycopg2==2.9.9
asyncpg==0.29.0
pytest-asyncio==0.24.0
jinja2==3.1.4 
bcrypt==4.2.0